refactor(share): turn the share screen into a bottom drawer
Was a full page navigation (router.push to /share/[id]) — jarring for
something that's really just "pick how to share this". Matches the
event-sharing UX in hemocione-digital-event (ElDrawer, direction="btt")
instead: same request card preview, same Stories/WhatsApp/Mais options,
opened straight from the description page without leaving it.

The shared URL was always /description/${id}, never /share/${id} — so
this loses no SEO surface, description/[id].vue already owns the real
OG tags for the link recipients land on.
